fixing creatures view => now we have a nice error message displayed if something bad happen

+++ src/views/creatures.hpp
@@ -29,7 +29,7 @@ private:
     sf::Text cnam, clev, ctyp, csta, catk, cdef, clif;
     sf::Sprite cimg;
     sf::Text error_msg;
-    int err_duration;
+    float err_duration;
 
     int selected;
     int offset;
